+++ custom_components/goecharger/config_flow.py
@@ -6,7 +6,7 @@ from homeassistant import config_entries
 from homeassistant.core import callback
 
 from homeassistant.const import CONF_HOST, CONF_SCAN_INTERVAL
-from .const import DOMAIN, CONF_NAME, CONF_CORRECTION_FACTOR
+from .const import API_VERSIONS, CONF_API_VERSION, DEFAULT_API_VERSION, DOMAIN, CONF_NAME, CONF_CORRECTION_FACTOR
 _LOGGER = logging.getLogger(__name__)
 
 
@@ -23,7 +23,7 @@ class ConfigFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
     async def async_get_options_flow(config_entry):
         return OptionsFlowHandler(config_entry)
 
-    async def async_step_user(self, info):
+    async def async_step_user(self, info=None):
         if info is not None:
             _LOGGER.debug(info)
             return self.async_create_entry(title=info[CONF_NAME], data=info)
@@ -39,6 +39,40 @@ class ConfigFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
                     vol.Required(
                         CONF_CORRECTION_FACTOR, default="1.0"
                     ): str,
+                    vol.Optional(
+                        CONF_API_VERSION, default=DEFAULT_API_VERSION
+                    ): vol.In(API_VERSIONS),
+                }
+            ),
+        )
+
+
+class OptionsFlowHandler(config_entries.OptionsFlow):
+    def __init__(self, config_entry):
+        self.config_entry = config_entry
+
+    async def async_step_init(self, user_input=None):
+        if user_input is not None:
+            return self.async_create_entry(title="", data=user_input)
+
+        api_version = self.config_entry.options.get(
+            CONF_API_VERSION,
+            self.config_entry.data.get(CONF_API_VERSION, DEFAULT_API_VERSION),
+        )
+        scan_interval = self.config_entry.options.get(
+            CONF_SCAN_INTERVAL, self.config_entry.data.get(CONF_SCAN_INTERVAL, 20)
+        )
+        correction_factor = self.config_entry.options.get(
+            CONF_CORRECTION_FACTOR,
+            self.config_entry.data.get(CONF_CORRECTION_FACTOR, "1.0"),
+        )
+        return self.async_show_form(
+            step_id="init",
+            data_schema=vol.Schema(
+                {
+                    vol.Optional(CONF_API_VERSION, default=api_version): vol.In(API_VERSIONS),
+                    vol.Optional(CONF_SCAN_INTERVAL, default=scan_interval): int,
+                    vol.Optional(CONF_CORRECTION_FACTOR, default=correction_factor): str,
                 }
             ),
         )