+++ custom_components/goecharger/switch.py
@@ -1,13 +1,11 @@
 """Platform for go-eCharger switch integration."""
 import logging
 from homeassistant.components.switch import SwitchEntity
-from homeassistant.const import CONF_HOST
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant import core, config_entries
 
-from goecharger import GoeCharger
-
 from .const import DOMAIN, CONF_CHARGERS, CONF_NAME, CHARGER_API, charger_entity_id
+from .api import GoeChargerV2, V2_SWITCHES
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -22,25 +20,9 @@ async def async_setup_entry(
     config = config_entry.as_dict()["data"]
 
     chargerName = config[CONF_NAME]
-    host = config[CONF_HOST]
-    chargerApi = GoeCharger(host)
+    chargerApi = hass.data[DOMAIN]["api"][chargerName]
 
-    entities = []
-
-    attribute = "allow_charging"
-    entities.append(
-        GoeChargerSwitch(
-            hass.data[DOMAIN]["coordinator"],
-            hass,
-            chargerApi,
-            charger_entity_id("switch", chargerName, attribute),
-            chargerName,
-            "Charging allowed",
-            attribute,
-        )
-    )
-
-    async_add_entities(entities)
+    async_add_entities(_create_switches_for_charger(hass, chargerName, chargerApi))
 
 
 async def async_setup_platform(hass, config, async_add_entities, discovery_info=None):
@@ -57,29 +39,49 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
     for charger in chargers:
         chargerName = charger[0][CONF_NAME]
 
-        attribute = "allow_charging"
-        entities.append(
-            GoeChargerSwitch(
-                hass.data[DOMAIN]["coordinator"],
-                hass,
-                chargerApi[chargerName],
-                charger_entity_id("switch", chargerName, attribute),
-                chargerName,
-                "Charging allowed",
-                attribute,
-            )
-        )
+        entities.extend(_create_switches_for_charger(hass, chargerName, chargerApi[chargerName]))
     async_add_entities(entities)
 
 
+def _create_switches_for_charger(hass, chargerName, chargerApi):
+    switches = [
+        GoeChargerSwitch(
+            hass.data[DOMAIN]["coordinator"],
+            hass,
+            chargerApi,
+            charger_entity_id("switch", chargerName, "allow_charging"),
+            chargerName,
+            "Charging allowed",
+            "allow_charging",
+        )
+    ]
+    if isinstance(chargerApi, GoeChargerV2):
+        for key, description in V2_SWITCHES.items():
+            attribute = description["attribute"]
+            switches.append(
+                GoeChargerSwitch(
+                    hass.data[DOMAIN]["coordinator"],
+                    hass,
+                    chargerApi,
+                    charger_entity_id("switch", chargerName, attribute),
+                    chargerName,
+                    description["name"],
+                    attribute,
+                    key,
+                )
+            )
+    return switches
+
+
 class GoeChargerSwitch(CoordinatorEntity, SwitchEntity):
-    def __init__(self, coordinator, hass, goeCharger, entity_id, chargerName, name, attribute):
+    def __init__(self, coordinator, hass, goeCharger, entity_id, chargerName, name, attribute, api_key=None):
         """Initialize the go-eCharger switch."""
         super().__init__(coordinator)
         self.entity_id = entity_id
         self._chargername = chargerName
         self._name = name
         self._attribute = attribute
+        self._api_key = api_key
         self.hass = hass
         self._goeCharger = goeCharger
         self._state = None
@@ -98,18 +100,24 @@ class GoeChargerSwitch(CoordinatorEntity, SwitchEntity):
 
     async def async_turn_on(self, **kwargs):
         """Turn the entity on."""
-        await self.hass.async_add_executor_job(self._goeCharger.setAllowCharging, True)
+        if self._api_key:
+            await self.hass.async_add_executor_job(self._goeCharger.setApiKey, self._api_key, True)
+        else:
+            await self.hass.async_add_executor_job(self._goeCharger.setAllowCharging, True)
         await self.coordinator.async_request_refresh()
 
     async def async_turn_off(self, **kwargs):
         """Turn the entity off."""
-        await self.hass.async_add_executor_job(self._goeCharger.setAllowCharging, False)
+        if self._api_key:
+            await self.hass.async_add_executor_job(self._goeCharger.setApiKey, self._api_key, False)
+        else:
+            await self.hass.async_add_executor_job(self._goeCharger.setAllowCharging, False)
         await self.coordinator.async_request_refresh()
 
     @property
     def name(self):
         """Return the name of the switch."""
-        return self._chargername
+        return self._name
 
     @property
     def unique_id(self):